refactor: Rename file from mapView.tsx to MapView.tsx
